refactor translator
