+++ packages/api/src/core/block/question.ts
@@ -1,10 +1,11 @@
-import { map, uniqBy } from 'lodash'
-import { Block, getPlainTextFromTokens } from '.'
-import { BlockType } from '../../types/block'
-import { LinkType } from '../../types/link'
-import { Token } from '../../types/token'
-import { Link } from '../link'
-import { extractPartialQueries } from '../translator/question'
+import { map, uniqBy } from 'lodash';
+
+import { Block, getPlainTextFromTokens } from '.';
+import { BlockType } from '../../types/block';
+import { LinkType } from '../../types/link';
+import { Token } from '../../types/token';
+import { Link } from '../link';
+import { extractPartialQueries } from '../translator';
 
 type QuestionBlockContent = {
   title: Token[]
