+++ packages/api/src/core/translator/interface.ts
@@ -1,3 +1,5 @@
+import { Block } from '../block'
+
 /**
  * Specific SQL actuator. Responsible to escape and perform SQL
  */
@@ -5,10 +7,10 @@ export interface ISqlTranslator {
   /**
    * Can SQL can be translated by this Translator
    */
-  match(param: string): boolean
+  match(block: Block): boolean
 
   /**
    * Assist param to an executable statement
    */
-  translate(param: string): Promise<string>
+  translate(block: Block): string
 }
